fix: multiple bumps on existing create-branch
Originally the rebase was only intended to update an existing branch
with the target specified in the `rebase-onto` option.  This action
would result in multiple bumps by release-it as it would see
previous increments in past commits.

Now, we only rebase-onto if the context branch and the create-branch
match.  Otherwise, if the `create-branch` option is specified,
we always create a new local branch off of the context branch.
This ensures the branch is always up to date.

Next, if the remote create-branch branch exists, we disable the
release-it `git.push` option and run release-it.  Then we rebase onto
the remote `create-branch`, skipping empty commits, setting the
local branch upstream to the remote branch, and force-push it.